[Core] Move ray-specific WorkerWrapperBase methods to RayWorkerWrapper
Noticed this while working on other items.
